+++ src/bda/plone/discount/browser/actions.py
@@ -1,0 +1,34 @@
+# -*- coding: utf-8 -*-
+from Products.Five.browser import BrowserView
+from bda.plone.discount.interfaces import IDiscountSettingsEnabled
+from zope.i18nmessageid import MessageFactory
+from zope.interface import directlyProvides
+from zope.interface import noLongerProvides
+
+
+_ = MessageFactory('bda.plone.discount')
+
+
+class EnableDisableDiscountAction(BrowserView):
+
+    def enable_discount(self):
+        directlyProvides(self.context, IDiscountSettingsEnabled)
+        self.context.portal_catalog.reindexObject(
+            self.context, idxs=['object_provides'], update_metadata=1)
+        self.context.plone_utils.addPortalMessage(
+            _(u'enabled_discount', default=u'Enabled discount on context.'))
+        self.request.response.redirect(self.context.absolute_url())
+
+    def disable_discount(self):
+        noLongerProvides(self.context, IDiscountSettingsEnabled)
+        self.context.portal_catalog.reindexObject(
+            self.context, idxs=['object_provides'], update_metadata=1)
+        self.context.plone_utils.addPortalMessage(
+            _(u'disabled_discount', default=u'Disabled discount on context.'))
+        self.request.response.redirect(self.context.absolute_url())
+
+    def can_enable_discount(self):
+        return not IDiscountSettingsEnabled.providedBy(self.context)
+
+    def can_disable_discount(self):
+        return IDiscountSettingsEnabled.providedBy(self.context)